+++ ExilenceNextApp/src/config/app.config.dev.ts
@@ -8,7 +8,8 @@ const devConfig = {
   oauthUrl: 'https://www.pathofexile.com',
   pathOfExileUrl: 'https://www.pathofexile.com',
   pathOfExileApiUrl: 'https://api.pathofexile.com',
-  pathOfExileCookieDomain: '.pathofexile.com'
+  pathOfExileCookieDomain: '.pathofexile.com',
+  poeNinjaBaseUrl: 'https://poe.ninja'
 };
 
 export default devConfig;