+++ ExilenceNextApp/src/interfaces/table-item.interface.ts
@@ -26,4 +26,5 @@ export interface ITableItem {
   tier: number;
   inventoryId: string;
   tabNames: string;
+  detailsUrl: string;
 }